pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "metalstats"
dynamic = ["version"]
description = 'API that turns your Spotify top tracks into a grid of album covers.'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "beucismis", email = "beucismis@tutamail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["uvicorn", "fastapi", "spotipy", "pillow", "requests", "pydantic-settings"]

[project.urls]
Source = "https://github.com/beucismis/metalstats"
Issues = "https://github.com/beucismis/metalstats/issues"
Documentation = "https://github.com/beucismis/metalstats#readme"

[tool.hatch.version]
path = "src/metalstats/__about__.py"

[tool.hatch.envs.default.scripts]
metalstats = "uvicorn src.metalstats.main:app --reload"
